fix : OPS-91 BigDecimal 대신 Long 반환하도록 수정

+++ src/main/java/com/example/tiggle/dto/donation/response/DonationStatus.java
@@ -1,9 +1,7 @@
 package com.example.tiggle.dto.donation.response;
 
-import java.math.BigDecimal;
-
 public record DonationStatus(
-        BigDecimal planetAmount,
-        BigDecimal peopleAmount,
-        BigDecimal prosperityAmount
+        Long planetAmount,
+        Long peopleAmount,
+        Long prosperityAmount
 ) {}
